refactor: split NextstrainGenbankService into download and database insertion
This makes it easier to fill the database with a predefined test data set.
